implement API to specify metadata request timeout
This commit adds a new API function to the driver:
`cass_cluster_set_metadata_request_serverside_timeout`.
This function allows users to set a server-side timeout for metadata
queries, which is particularly useful for large schemas where the
default timeout may be too tight. The default timeout is set to
2 seconds (as in Rust Driver), and users can pass 0 to use the cluster's
default timeout. This simply exposes the existing Rust Driver's API.

Fixes: #172