session: avoid `unwrap()`s in favour of let-else
This is a refactor to avoid using `unwrap()` on the `connected` field
of the `SessionGuard`. Instead, it uses the `let-else` syntax to handle
the case where the session is not connected in a more idiomatic way.

Note that no `unwrap()` could really panic here, as the `connected`
field was always checked before, but this change makes the code
safer and more readable.