future: put err_string under OnceLock
The `err_string` field in `CassFuture` serves as an allocation cache
for the error message string, which is generated only when needed
and must have a lifetime of `CassFuture`, as required by the
`cass_future_error_message` function. This field
was previously an `Option<String>` and was held under a `Mutex` inside
`CassFutureState`, but as its semantics fit `OnceLock` perfectlym it has
been changed accordingly and is now a `OnceLock<String>`.

As a pleasant result, `with_waited_result` can now be used instead of
`with_waited_state`, which simplifies the code and avoids `unwrap`.