future: ditch closure-based awaiting
There has long existed an API for awaiting futures that takes a closure
and invokes it with either the future state or the future result once
the future is ready.
This API is not really ergonomic, but it was a must as long as we were
holding the future result inside the future state, which is held behind
a mutex. As the result could be accessible only while the state's
`MutexGuard` is alive, we could not return references to the result
outside of the closure.
Now that we have the future result stored in an `OnceLock`, we can
return references to it, which makes the API much more ergonomic.
The future state is still held behind a mutex, so it's still not
possible to return a reference to the future state outside of
the closure, but after both the future result and the cached error
string were moved to `OnceLock`s out of the future state, we can
remove the closure-based future state awaiting API altogether,
and make the future result awaiting API reference-based instead
of a closure-based one. Perfect! Clean! Nice!

Refactors induced by this change in `waited_result(_timed)` has showed
that the logic there is a bit convoluted, so I added FIXMEs there.
I'm going to address them in further commits.