future: waiting logic readability improvements
`CassFuture::waited_result(_timed)` methods have been documented better
and refactored to improve readability.
**Logic has NOT been changed.**

Code changes & rationale:
The previous logic was a bit convoluted: it `unwrap`ped the result
immediately after both branches of the `if` related to `join_handle`
merged back, which made it hard to understand the flow of the code.
Now, the logic is clearer: the result is `unwrap`ped in the branch
where we are sure that the future is resolved, i.e. the `join_handle`
had been `Some` and we have just completed the future.
The other branch, where we wait for the future to be completed,
is now clearer as well: once we wake up, there are two cases (quoting
the new comments):
1. The result is already available, so we can return it.
2. The `join_handle` was `None`, and now it's `Some` - some
   other thread must have timed out and returned the handle.
   We need to take over the work of completing the future, because
   the result is still not available and we may be using
   `current_thread` tokio executor, in which case no one else will
   complete the future, so it's our responsibility. In the next
   iteration we will land in the branch with `block_on` and complete
   the future.
In both cases we just continue the loop, so the code is simpler
and easier to read.