future: introduce FutureKind
When trying to introduce configurable runtime to the CassFuture
framework, I realized that the current implementation of CassFuture
is not flexible enough. Namely, those futures that are immediately
ready with the result (e.g. futures that are created with
`CassFuture::new_ready`) do not have access to the runtime on creation.
To provide better type safety, this kind of future was extracted as a
separate enum variant of `FutureKind`, with the other variant
representing futures that must be resolved by the tokio runtime.

The next commit demonstrates how extracting these two kinds of futures
into a separate enum allows us to implement the configurable runtime
in the CassFuture framework.