session: handle disconnection errors synchronously
When issuing an asynchronous operation on a session, such as executing
or preparing a statement, let's synchronously ensure that the session
is connected before proceeding. If the session is not connected, return
an error immediately instead of allowing the operation to continue and
asynchronously fail later.

This change is not much observable by a user, but it's necessary to
introduce the next change involving runtime configuration.