future: optimise `cass_future_error_message()`
The function always allocates a new string for the error message when
called the first time on a future in order to return a pointer to that
allocated string to the caller. This is unnecessary when the future's
error kind is already the string form. This commit performs a small
refactor that makes allocation avoided for those two cases with errors
given in a string form.
Credits to @lorak-mmk for the suggestion.