future: fix setting callback to be at-most-once
CPP Driver has the semantics that a callback can be set only once for
the whole lifetime of a future. This commit fixes the Rust wrapper to
match this behavior. This is guaranteed by unconditionally setting
the callback field in the `set_callback` method, and by leaving the
callback field as `Some` after the callback is invoked in both cases:
1. when the future is already resolved, so the callback is invoked
   synchronously by the thread that sets the callback;
2. when the future is not resolved yet, so the callback is invoked
   asynchronously by the thread that resolves the future once it's
   resolved.