docs/security: extract nonsupported features
The documentation regarding custom authenticator methods, which are not
supported by the CPP-Rust Driver, has been extracted to a separate
file `nonsupported.md` in the `docs/source/topics/security/`
directory. This file is not part of the documentation tree, as we do not
want to advertise these features. It serves as a reference for
implementors who may want to implement these features in the future.